# backend/src/test/resources/openapi/api-docs.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Mail Backend API",
    "description" : "REST API for the Tauri desktop mail client — accounts, messages, drafts, folders.",
    "license" : {
      "name" : "Proprietary"
    },
    "version" : "1.0.0"
  },
  "security" : [ {
    "ApiKeyAuth" : [ ]
  } ],
...
  "paths" : {
    "/api/v1/accounts/{id}" : {
      "get" : {
        "tags" : [ "Accounts" ],
        "summary" : "Account detail",
        "description" : "Returns one specific e-mail account by ID.",
        "operationId" : "getAccountById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        } ],
        "responses" : {
          "503" : {
            "description" : "Service Unavailable"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Accounts" ],
        "summary" : "Full account update",
        "description" : "Replaces all account fields. Password/secret are re-encrypted if sent.",
        "operationId" : "updateAccount",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccountUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "503" : {
            "description" : "Service Unavailable"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input (validation error).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Accounts" ],
        "summary" : "Delete account",
        "description" : "Deletes the account and closes its IMAP connection (cleanup of the ImapConnectionManager pool).",
        "operationId" : "deleteAccount",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        } ],
        "responses" : {
          "503" : {
            "description" : "Service Unavailable"
          },
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/accounts" : {
      "get" : {
        "tags" : [ "Accounts" ],
        "summary" : "List all accounts",
        "description" : "Returns a list of all registered e-mail accounts (including inactive ones).",
        "operationId" : "getAllAccounts",
        "responses" : {
          "503" : {
            "description" : "Service Unavailable"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AccountResponse"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Accounts" ],
        "summary" : "Create account",
        "description" : "Creates a new account, encrypts the password or OAuth secret and returns 201 with the complete DTO.",
        "operationId" : "createAccount",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccountCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "503" : {
            "description" : "Service Unavailable"
          },
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input (validation error).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/accounts/test-connection" : {
      "post" : {
        "tags" : [ "Accounts" ],
        "summary" : "Account connection test",
        "description" : "Verifies IMAP and SMTP connections for the given provider/custom config without saving the account.",
        "operationId" : "testConnection",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccountConnectionTestRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "503" : {
            "description" : "Service Unavailable"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountConnectionTestResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input (validation error).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/system/readiness" : {
      "get" : {
        "tags" : [ "System" ],
        "summary" : "Returns backend readiness state",
        "description" : "The desktop client uses this response at startup. Readiness means the API is ready to serve the UI; the first mail synchronization and accounts waiting for re-auth do not impact readiness.",
        "operationId" : "getReadiness",
        "responses" : {
          "503" : {
            "description" : "Service Unavailable"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SystemReadinessResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ProblemDetail" : {
        "description" : "RFC 9457 problem+json — format of all error responses.",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "URI identifier of the error type (e.g. .../errors/account_not_found)."
          },
          "title" : {
            "type" : "string",
            "description" : "Short human-readable description of the error class."
          },
          "status" : {
            "description" : "HTTP status code."
          },
          "detail" : {
            "type" : "string",
            "description" : "Error detail intended for the user."
          },
          "instance" : {
            "type" : "string",
            "format" : "uri",
            "description" : "URI of the specific occurrence (typically the request URI)."
          },
          "errorCode" : {
            "type" : "string",
            "description" : "Machine-readable error code (e.g. ACCOUNT_NOT_FOUND, VALIDATION_ERROR)."
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Time when the error occurred (ISO-8601)."
          }
        }
      },
      "AccountUpdateRequest" : {
        "type" : "object",
        "properties" : {
          "accountName" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 255,
            "minLength" : 0
          },
          "displayName" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "providerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "imap" : {
            "$ref" : "#/components/schemas/MailServerSettings"
          },
          "smtp" : {
            "$ref" : "#/components/schemas/MailServerSettings"
          },
          "username" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "password" : {
            "type" : "string"
          },
          "active" : {
            "type" : "boolean"
          },
          "providerOrCustomServerConfigPresent" : {
            "type" : "boolean"
          }
        },
        "required" : [ "accountName", "email", "username" ]
      },
      "MailServerSettings" : {
        "type" : "object",
        "properties" : {
          "host" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "port" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 65535,
            "minimum" : 1
          },
          "useSsl" : {
            "type" : "boolean"
          }
        },
        "required" : [ "host", "port", "useSsl" ]
      },
      "AccountResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "accountName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "providerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "providerName" : {
            "type" : "string"
          },
          "imapHost" : {
            "type" : "string"
          },
          "imapPort" : {
            "type" : "integer",
            "format" : "int32"
          },
          "imapUseSsl" : {
            "type" : "boolean"
          },
          "smtpHost" : {
            "type" : "string"
          },
          "smtpPort" : {
            "type" : "integer",
            "format" : "int32"
          },
          "smtpUseSsl" : {
            "type" : "boolean"
          },
          "username" : {
            "type" : "string"
          },
          "authType" : {
            "type" : "string",
            "enum" : [ "PASSWORD", "OAUTH2" ]
          },
          "oauth2Provider" : {
            "type" : "string"
          },
          "active" : {
            "type" : "boolean"
          },
          "requiresReauth" : {
            "type" : "boolean"
          },
          "lastSyncAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastError" : {
            "type" : "string"
          },
          "lastErrorCode" : {
            "type" : "string"
          },
          "lastErrorArgs" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        }
      },
...
      "AccountCreateRequest" : {
        "type" : "object",
        "properties" : {
          "accountName" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "displayName" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 255,
            "minLength" : 0
          },
          "providerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "imap" : {
            "$ref" : "#/components/schemas/MailServerSettings"
          },
          "smtp" : {
            "$ref" : "#/components/schemas/MailServerSettings"
          },
          "username" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "password" : {
            "type" : "string",
            "minLength" : 1
          },
          "providerOrCustomServerConfigPresent" : {
            "type" : "boolean"
          }
        },
        "required" : [ "accountName", "email", "password", "username" ]
      },
...
      "AccountConnectionTestRequest" : {
        "type" : "object",
        "properties" : {
          "accountId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 255,
            "minLength" : 0
          },
          "providerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "imap" : {
            "$ref" : "#/components/schemas/MailServerSettings"
          },
          "smtp" : {
            "$ref" : "#/components/schemas/MailServerSettings"
          },
          "username" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "password" : {
            "type" : "string"
          },
          "providerOrCustomServerConfigPresent" : {
            "type" : "boolean"
          },
          "passwordPresentForNewAccount" : {
            "type" : "boolean"
          }
        },
        "required" : [ "email", "username" ]
      },
      "AccountConnectionTestResponse" : {
        "type" : "object",
        "properties" : {
          "imapOk" : {
            "type" : "boolean"
          },
          "smtpOk" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "SystemReadinessResponse" : {
        "type" : "object",
        "properties" : {
          "ready" : {
            "type" : "boolean",
            "description" : "True if the backend can serve user-facing API requests.",
            "example" : true
          },
          "phase" : {
            "type" : "string",
            "description" : "Current phase of the readiness state machine.",
            "enum" : [ "READY" ],
            "example" : "READY"
          },
          "appName" : {
            "type" : "string",
            "description" : "Backend application name.",
            "example" : "mail"
          },
          "appVersion" : {
            "type" : "string",
            "description" : "Backend application version.",
            "example" : "0.1.0"
          },
          "apiVersion" : {
            "type" : "string",
            "description" : "API contract version.",
            "example" : "1.0.0"
          },
          "minClientVersion" : {
            "type" : "string",
            "description" : "Minimum supported frontend version.",
            "example" : "0.0.1"
          },
          "dbSchemaVersion" : {
            "type" : "string",
            "description" : "Current Flyway DB schema version.",
            "example" : "1"
          },
          "reason" : {
            "type" : "string",
            "description" : "Safe user-facing reason when ready=false."
          }
        }
      },
...
    },
    "securitySchemes" : {
      "ApiKeyAuth" : {
        "type" : "apiKey",
        "description" : "Internal API key that the desktop client loads from session.json and sends in the X-API-KEY header.",
        "name" : "X-API-KEY",
        "in" : "header"
      }
    }
  }
}